testutils: move testutils package to internal/
