+++ resolve/glob/glob_test.go
@@ -6,8 +6,8 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/simplesurance/baur/testutils/fstest"
-	"github.com/simplesurance/baur/testutils/strtest"
+	"github.com/simplesurance/baur/internal/testutils/fstest"
+	"github.com/simplesurance/baur/internal/testutils/strtest"
 )
 
 func Test_FindAllSubDirs(t *testing.T) {
