+++ storage/postgres/init_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/simplesurance/baur/testutils/dbtest"
+	"github.com/simplesurance/baur/internal/testutils/dbtest"
 )
 
 var ctx = context.Background()
